Move AccountSelect values to MobX store
